Decouple leave match modal
